same

+++ RRRPG-main/RRRPG-main/RRRPGLib/multiPlayer.cs
@@ -376,7 +376,7 @@ namespace RRRPGLib
         {
             if (name == "")
                 name = this.id.ToString();
-            sendMessage(this.id.ToString() + (char)127 + character.ToString() + (char)127 + name);
+            sendMessage(this.id.ToString() + (char)127 + ((int)character).ToString() + (char)127 + name);
         }
     }
 }
